MeasureX
1. “Load Calibration” for session removed from tool bar and placed into “Session” properties screen as a button on the bottom of it.
Reviewed enabled/disabled state of “View” icons. “Mask Editor”, “Interpolation” and so fourth should not be available if we do not have a valid session.

2. Chart. “Average Pressure” title appears in two places, removed one of them.
pressure units, like PSI in calibrated mode are clearly displayed

3. Histogram. pressure units, like PSI in calibrated mode are clearly displayed
“Histogram” title appears in three places, remove two of them.
Frequency” label should be substituted for something else
After the only session is removed the histogram is “STILL” displayed.

4. Status bar. pressure units, like PSI in calibrated mode are clearly displayed

5. Device Settings Screen: Changed title. Added more space for the calibration name.