Calibrator: 1. Now it is able to load previously saved calibrations
Mex:
> 4)      Numbers shown in status bar should all be formatted
I found only "Sensel max" and "Sensel min" unformatted. Fixed.

> 5)      When uS units are used status bar statistics do  not refresh as the cursor is moving around
fixed

> 7)      When alt clicking the 2D cell we get a new mask, but it is not displayed on the chart and statistics for it are not calculated. After it shows on the screen, alt clicking on other cells does not produce proper charts also.
fixed

> 8)      Why turning off calibrated mode on a large file is taking that long? We already have all the calibrated data.
fixed

> 10)   Is the moving average working correctly.
Fixed. The performance of "Moving average" is disgusting and needs to be optimized.

> 12)    After closing Player Settings there is a very large delay
Delay depends on the calibration mode and Moving average settings which are selected in Player settings. Also the number of frames that are shown in the chart painter is very significant. When calibration and moving average are turned off, the delay is now minimal. When they are turned off the special algorithms for the partial drawing of the chart must be developed.

> 13)   “Numbers” session was already opened.  Without any device tried to record with a hand plate driver. Clicked cancel instead of searching a port. Application seemed hang for several minutes but after that MeX continued to work. Device settings icon on the menu and toolbar seemed desynchronized from enabled/disabled point of view.
Not implemented yet, since requires some redesign, Now a reference to a session that is being recorded is returned immediately from DeviceReader.record() method. However, handling of the device errors is performed in the other thread.