MeX:
For DEBUG constant unchecked in MeXCore and MeasureX projects the following functionality is disabled:
1. Total pressure is invisible in PainterFrameInfo
2. "Use calibration" button is invisible on the tool bar
3. Label with text "Calibration on" is always invisible
Nevertheless, it is still possible to turon off the calibration via "Edit Properties" windows (either device or session)

OrthoScan:
Some changes to show the history of the centers of gravity on a composite frame. It is now added at the end of a session.